Complete multisite tests and integrate in CI
