+++ .github/workflows/ci.yml
@@ -62,3 +62,4 @@ jobs:
           sudo /etc/init.d/mysql start
           bash bin/install-wp-tests.sh wordpress_test root 'root' localhost latest
           vendor/bin/phpunit
+          vendor/bin/phpunit -c tests/phpunit/multisite.xml
